+++ core/chains/evm/bulletprooftxmanager/eth_confirmer.go
@@ -380,13 +380,13 @@ func (ec *EthConfirmer) getNonceForLatestBlock(ctx context.Context, from gethCom
 
 // Note this function will increment promRevertedTxCount upon receiving
 // a reverted transaction receipt. Should only be called with unconfirmed attempts.
-func (ec *EthConfirmer) batchFetchReceipts(ctx context.Context, attempts []EthTxAttempt) (receipts []Receipt, err error) {
+func (ec *EthConfirmer) batchFetchReceipts(ctx context.Context, attempts []EthTxAttempt) (receipts []evmtypes.Receipt, err error) {
 	var reqs []rpc.BatchElem
 	for _, attempt := range attempts {
 		req := rpc.BatchElem{
 			Method: "eth_getTransactionReceipt",
 			Args:   []interface{}{attempt.Hash},
-			Result: &Receipt{},
+			Result: &evmtypes.Receipt{},
 		}
 		reqs = append(reqs, req)
 	}
@@ -402,9 +402,9 @@ func (ec *EthConfirmer) batchFetchReceipts(ctx context.Context, attempts []EthTx
 		attempt := attempts[i]
 		result, err := req.Result, req.Error
 
-		receipt, is := result.(*Receipt)
+		receipt, is := result.(*evmtypes.Receipt)
 		if !is {
-			return nil, errors.Errorf("expected result to be a %T, got %T", (*Receipt)(nil), result)
+			return nil, errors.Errorf("expected result to be a %T, got %T", (*evmtypes.Receipt)(nil), result)
 		}
 
 		l := lggr.With(
@@ -461,7 +461,7 @@ func (ec *EthConfirmer) batchFetchReceipts(ctx context.Context, attempts []EthTx
 	return
 }
 
-func (ec *EthConfirmer) saveFetchedReceipts(receipts []Receipt) (err error) {
+func (ec *EthConfirmer) saveFetchedReceipts(receipts []evmtypes.Receipt) (err error) {
 	if len(receipts) == 0 {
 		return nil
 	}
